requirements.txt
streamlit
pandas
pdfplumber
python-docx
scikit-learn
spacy
nltk
google-generativeai
python-dotenv
docx2txt  # Add this line